==> scripts/1.merging_met_filesBB2.Rmd <==
---
title: "1.merging_met_files"
author: "Sara Knox; revised Marion Nyberg"
date: "04/09/2019/; revised 14/05/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())

# load libraries
library(dplyr)
library(gtools)
library(gdata)
library(ggplot2)
library(tidyr)
library(plotly)

# set wd
getwd()
#dir <- "../Flux-tower/met_data/cr1000"
dir <- "G:/.shortcut-targets-by-id/1txCh9lZ7VGCujXGvBaJCMVnuxT-65q4K/Micromet Lab/Projects/2019-BB2 Burns Bog 2/Flux-tower (1)" #Marion's wd
knitr::opts_knit$set(root.dir = paste0(dir, "/met_data/cr1000"))
```

```{r load data, echo=FALSE, include=FALSE}

# Identify files
raw.files <- Sys.glob("*.csv")

# Initialize with Air temperature file - longest data record
cr1000 <- read.csv("BB2DTA.csv",
                   header = TRUE,
                   sep = ",")
colnames(cr1000)[1] <- "Timestamp"
cr1000$Timestamp <- as.POSIXct(cr1000$Timestamp, format = "%Y-%m-%d %H:%M")

# Remova NA values (coinciding with daylight savings)
cr1000 <- cr1000[!is.na(cr1000$Timestamp), ]

# Remove BB2 column and extra X column (not sure where this came from)
cr1000 <- cr1000[,(-grep("X", colnames(cr1000)))]

Length <- nrow(cr1000)


# Now add other variable to dataframe
for (i in raw.files[2:length(raw.files)]) {
  sample <- read.csv(i,
                     header = TRUE,
                     sep = ",")
  colnames(sample)[1] <- "Timestamp"
  sample$Timestamp <- as.POSIXct(sample$Timestamp, format = "%Y-%m-%d %H:%M")
  
  # Remova NA values (coinciding with daylight savings)
  sample <- sample[!is.na(sample$Timestamp), ]
  
  # Make all imported files the same length
  sample <- slice(sample, 1: Length)

  cr1000 <-merge(x=cr1000,y=sample,by="Timestamp",all=TRUE)
}

cr1000$Timestamp <- as.POSIXct(cr1000$Timestamp, format = "%Y-%m-%d %H:%M")

cr1000 <- cr1000[,(-grep("X", colnames(cr1000)))]

names(cr1000)

# If adding a new variable, rename here!
#names(cr1000) <- c("Timestamp", "AIR_TEMP_2M", "AIR_TEMP_38CM",
                   #"ORP_10CM", "ORP_30CM", "INCOMING_PAR", "REFLECTED_PAR",
                   #"PRECIP", "PA_EC1", "PA_EC2_TOTAL","PA_EC2_AIR","PA_2M","SHORTWAVE_IN", "SHORTWAVE_OUT",
                   #"LONGWAVE_IN", "LONGWAVE_OUT", "RH_2M", "RH_38CM",
                  # "SHFP_1", "SHFP_2", "SHFP_3", "SVWC", "SOIL_TEMP_5CM",
                   #"SOIL_TEMP_10CM", "SOIL_TEMP_50CM", "WTH", "BOG_HEIGHT", "WATER_TEMP_10CM",
                   #"WATER_TEMP_30CM", "WIND_VELOCITY_CUP", "WIND_VELOCITY_EC1", "WIND_VELOCITY_EC2")

names(cr1000) <- c("Timestamp", "AIR_TEMP_2M", "INCOMING_PAR", "REFLECTED_PAR",
                  "PRECIP","PA_EC_AIR2.5M","PA_1.5M","SHORTWAVE_IN", "SHORTWAVE_OUT",
                   "LONGWAVE_IN", "LONGWAVE_OUT", "RH_2M",
                    "SHFP_1", "SHFP_2", "SHFP_3", "SVWC", "SOIL_TEMP_1_5CM",
                  "SOIL_TEMP_1_10CM", "SOIL_TEMP_1_30CM", "SOIL_TEMP_1_50CM","SOIL_TEMP_2_5CM",
                   "SOIL_TEMP_2_10CM","SOIL_TEMP_2_30CM", "SOIL_TEMP_2_50CM","SOIL_TEMP_3_5CM",
                  "SOIL_TEMP_3_10CM","SOIL_TEMP_3_30CM", "SOIL_TEMP_3_50CM", "WTH","WATER_TEMP_1_5CM","WATER_TEMP_2_5CM","WATER_TEMP_3_5CM", "WIND_VELOCITY_CUP", "WIND_VELOCITY_EC")

#names(cr1000) <- c("Timestamp", "AIR_TEMP_2M", "INCOMING_PAR", "REFLECTED_PAR",
 #                  "PRECIP","PA_EC_AIR2.5M","PA_1.5M","SHORTWAVE_IN", "SHORTWAVE_OUT",
  #                 "LONGWAVE_IN", "LONGWAVE_OUT", "RH_2M",
   #                 "SHFP_1", "SHFP_2", "SHFP_3", "SVWC", "SOIL_TEMP_1_5CM",
    #               "SOIL_TEMP_1_10CM", "SOIL_TEMP_1_30CM", "SOIL_TEMP_1_50CM","SOIL_TEMP_2_5CM",
     #              "SOIL_TEMP_2_10CM","SOIL_TEMP_2_30CM", "SOIL_TEMP_2_50CM","SOIL_TEMP_3_5CM",
      #             "SOIL_TEMP_3_10CM","SOIL_TEMP_3_30CM", "SOIL_TEMP_3_50CM", "WTH", "WIND_VELOCITY_CUP", "WIND_VELOCITY_EC")

# Convert all NaN to NA
is.nan.data.frame <- function(cr1000)
  do.call(cbind, lapply(cr1000, is.nan))

cr1000[is.nan(cr1000)] <- NA

#cr1000 <- cr1000[, -grep("WIND_VELOCITY_EC1", colnames(cr1000))]
#cr1000 <- cr1000[, -grep("PA_EC2_TOTAL", colnames(cr1000))]
#cr1000 <- cr1000[, -grep("WIND_VELOCITY_EC2", colnames(cr1000))]

head(cr1000)
str(cr1000)

# Create year & DOY variables
cr1000 <- cr1000 %>%
  mutate(year = as.numeric(format(Timestamp,'%Y')),
         DOY = as.numeric(format(Timestamp,'%j')))

cr1000$year <- factor(cr1000$year)
```

```{r plot air temperature, echo=FALSE, message=FALSE}

# Air temperature - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~AIR_TEMP_2M, name = '2M', type = 'scatter', mode = 'lines') 

# Air temperature - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = AIR_TEMP_2M, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```


```{r plot PAR, echo=FALSE, message=FALSE}
# PAR - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~INCOMING_PAR, name = 'Incoming', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~REFLECTED_PAR, name = 'Outgoing', mode = 'lines')

# PAR - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = INCOMING_PAR, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = REFLECTED_PAR, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot PA, echo=FALSE, message=FALSE}
# PA - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~PA_EC_AIR2.5M, name = 'PA_EC_AIR2.5M', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~PA_1.5M, name = 'PA 1.5M', mode = 'lines')
```

```{r plot PRECIP, echo=FALSE, message=FALSE}
# PRECIP - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~PRECIP, type = 'scatter', mode = 'lines') 

# PRECIP - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = PRECIP, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot VWC, echo=FALSE, message=FALSE}
# SVWC - full time series - look into further using Johannes' code!
plot_ly(data = cr1000, x = ~Timestamp, y = ~SVWC, type = 'scatter', mode = 'lines') 

# SVWC - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = SVWC, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot WTH, echo=FALSE, message=FALSE}
# WTH - full time series - look into further using Johannes' code!
plot_ly(data = cr1000, x = ~Timestamp, y = ~WTH, type = 'scatter', mode = 'lines') 

# WTH - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = WTH, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```


```{r plot SW, echo=FALSE, message=FALSE}
# SW - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~SHORTWAVE_IN, name = 'Incoming', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SHORTWAVE_OUT, name = 'Outgoing', mode = 'lines')

# SW - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = SHORTWAVE_IN, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SHORTWAVE_OUT, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot LW, echo=FALSE, message=FALSE}
# LW - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~LONGWAVE_IN, name = 'Incoming', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SHORTWAVE_OUT, name = 'LONGWAVE_OUT', mode = 'lines')

# LW - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = LONGWAVE_IN, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = LONGWAVE_OUT, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot RH, echo=FALSE, message=FALSE}
# RH - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~RH_2M, name = '2 m', type = 'scatter', mode = 'lines')

# RH - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = RH_2M, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

```

```{r plot SHFP, echo=FALSE, message=FALSE}
# SHFP - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~SHFP_1, name = 'SHFP 1', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SHFP_2, name = 'SHFP 2', mode = 'lines') %>%
  add_trace(y = ~SHFP_3, name = 'SHFP 3', mode = 'lines')

# SHFP - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = SHFP_1, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SHFP_2, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SHFP_3, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot TS, echo=FALSE, message=FALSE}
# TS - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~SOIL_TEMP_1_5CM, name = '1_5 cm', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_1_10CM, name = '1_10 cm', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_1_30CM, name = '1_30 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_1_50CM, name = '1_50 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_2_5CM, name = '2_10 cm', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_2_10CM, name = '2_30 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_2_30CM, name = '2_50 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_2_50CM, name = '2_50 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_3_5CM, name = '3_10 cm', mode = 'lines') %>%
  add_trace(y = ~SOIL_TEMP_3_10CM, name = '3_30 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_3_30CM, name = '3_50 cm', mode = 'lines') %>%
	add_trace(y = ~SOIL_TEMP_3_50CM, name = '3_50 cm', mode = 'lines')

# TS - by year - change the variables to view other soil temp reps
ggplot(data = cr1000, mapping = aes(x = DOY, y = SOIL_TEMP_1_5CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SOIL_TEMP_1_10CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SOIL_TEMP_1_30CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = SOIL_TEMP_1_50CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot TW, echo=FALSE, message=FALSE}
# WATER TEMP - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~WATER_TEMP_1_5CM, name = '1_5 cm', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~WATER_TEMP_2_5CM, name = '2_5 cm', mode = 'lines') %>%
	add_trace(y = ~WATER_TEMP_3_5CM, name = '3_5 cm', mode = 'lines')

# TW - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = WATER_TEMP_1_5CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = WATER_TEMP_2_5CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = WATER_TEMP_3_5CM, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r plot WS, echo=FALSE, message=FALSE}
# WS - full time series
plot_ly(data = cr1000, x = ~Timestamp, y = ~WIND_VELOCITY_CUP, name = 'Cup', type = 'scatter', mode = 'lines') %>%
  add_trace(y = ~WIND_VELOCITY_EC, name = 'EC', mode = 'lines')

# WS - by year
ggplot(data = cr1000, mapping = aes(x = DOY, y = WIND_VELOCITY_CUP, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))

ggplot(data = cr1000, mapping = aes(x = DOY, y = WIND_VELOCITY_EC, na.rm = TRUE)) +
  geom_line(aes(colour=year, group=year))
```

```{r save output, echo=FALSE, message=FALSE}
write.csv(cr1000, paste0(dir, '/met_data/met_merged/met_mergedBB2.csv'),row.names=FALSE)
```
